restructure log time tests
